fix: resolve clippy warnings across workspace
- Remove dead code: unused functions, `if false` branches, unused
  variables, unreachable expressions
- Remove `Slot::data_mut(&self)` returning `&mut [u8]` from immutable
  input; use `copy_nonoverlapping` via `data_ptr()` instead
- Replace manual `(x + y - 1) / y` with `.div_ceil()` in mlx5 WQE
  macros and scalerpc/erpc
- Replace manual clamp patterns with `.clamp()`
- Replace manual `% interval != 0` with `.is_multiple_of()`
- Convert needless range loops to iterators where possible
- Add `#[allow]` for needless_range_loop on dual-index matrix patterns
  and for too_many_arguments on protocol constructors
- Fix `to_*` methods on Copy types to take `self` by value
- Add type alias for complex `Vec<Vec<Option<Sender>>>` type

+++ scalerpc/src/protocol.rs
@@ -393,6 +393,7 @@ impl ResponseHeader {
     }
 
     /// Create a new response header with processing pool info AND credit acknowledgment.
+    #[allow(clippy::too_many_arguments)]
     pub fn with_processing_pool_info_and_credit_ack(
         req_id: u64,
         status: u32,
@@ -462,7 +463,7 @@ impl ResponseHeader {
     ///
     /// This is the lower 32 bits of context_switch_seq.
     pub fn get_context_switch_seq(&self) -> u64 {
-        (self.context_switch_seq & 0xFFFFFFFF) as u64
+        self.context_switch_seq & 0xFFFFFFFF
     }
 
     /// Get the fetched endpoint entry sequence (valid when has_context_switch() is true).
